# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing product details with clean architecture",
        "title": "Product Detail API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/products": {
            "get": {
                "description": "Get a list of all products with their details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get all products",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/desafio_mercado_livre_src_internal_entity.ProductDetail"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_infra_web_controller.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new product with all its details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Create a new product",
                "parameters": [
                    {
                        "description": "Product Details",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/desafio_mercado_livre_src_internal_entity.ProductDetail"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/desafio_mercado_livre_src_internal_entity.ProductDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_infra_web_controller.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_infra_web_controller.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "description": "Get detailed information about a specific product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get a product by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/desafio_mercado_livre_src_internal_entity.ProductDetail"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_infra_web_controller.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_infra_web_controller.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "desafio_mercado_livre_src_internal_entity.ProductDetail": {
            "type": "object",
            "required": [
                "brand",
                "category",
                "color",
                "description",
                "discount_percent",
                "free_shipping",
                "highlights",
                "images",
                "model",
                "name",
                "original_price",
                "payment_options",
                "price",
                "product_id",
                "purchase_options",
                "rating",
                "related_products",
                "review_count",
                "seller",
                "specs",
                "stock",
                "warranty"
            ],
            "properties": {
                "brand": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "discount_percent": {
                    "type": "number",
                    "maximum": 100,
                    "minimum": 0
                },
                "free_shipping": {
                    "type": "boolean"
                },
                "highlights": {
                    "description": "Lista de destaques do produto",
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "description": "ID interno auto-incrementado",
                    "type": "integer"
                },
                "images": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                },
                "model": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "original_price": {
                    "type": "number"
                },
                "payment_options": {
                    "description": "Lista de opções de pagamento aceitas",
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                },
                "price": {
                    "type": "number"
                },
                "product_id": {
                    "description": "ID externo do produto (ex: MLB123456)",
                    "type": "string"
                },
                "purchase_options": {
                    "description": "Opções de compra disponíveis",
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                },
                "rating": {
                    "description": "Avaliação média do produto",
                    "type": "number",
                    "maximum": 5,
                    "minimum": 0
                },
                "related_products": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "review_count": {
                    "type": "integer",
                    "minimum": 0
                },
                "seller": {
                    "$ref": "#/definitions/desafio_mercado_livre_src_internal_entity.SellerInfo"
                },
                "specs": {
                    "$ref": "#/definitions/desafio_mercado_livre_src_internal_entity.ProductSpecs"
                },
                "stock": {
                    "type": "integer",
                    "minimum": 0
                },
                "updated_at": {
                    "type": "string"
                },
                "warranty": {
                    "type": "string"
                }
            }
        },
        "desafio_mercado_livre_src_internal_entity.ProductSpecs": {
            "type": "object",
            "required": [
                "attributes"
            ],
            "properties": {
                "attributes": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "desafio_mercado_livre_src_internal_entity.SellerInfo": {
            "type": "object",
            "required": [
                "name",
                "official",
                "reputation",
                "sales",
                "type"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "official": {
                    "type": "boolean"
                },
                "reputation": {
                    "type": "string"
                },
                "sales": {
                    "type": "integer",
                    "minimum": 0
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "internal_infra_web_controller.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Error message"
                }
            }
        }
    }
}